Protected routes with authentication middleware

+++ api/app.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const app = express();
+const jwt = require('jsonwebtoken')
 
 const {mongoose} = require('./db/mongoose');
 
@@ -26,6 +27,24 @@ app.use(function (req, res, next) {
 
     next();
 });
+
+// Check whether the request has a valid JWT token
+let authenticate = (req, res, next ) => {
+    let token = req.header('x-access-token');
+
+    //verify JWT token
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if (err) {
+            // there was an error
+            // jwt is invalid
+            res.status(401).send(err);
+        } else {
+            // jwt is valid
+            req.user_id = decoded._id;
+            next()
+        }
+    })
+}
 
 
 // Verify refresh token middleware (which will be verifying the session)
@@ -81,19 +100,24 @@ let verifySession = (req, res, next) => {
 // List Routes
 
 // Get all lists
-app.get('/lists', (req, res) => {
-    // Return an array of all the lists in the database
-    List.find({}).then((lists) => {
+app.get('/lists', authenticate, (req, res) => {
+    // Return an array of all the lists in the database belongs to the authenticated user
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
+    }).catch((e) => {
+        res.send(e)
     });
 });
 
 // Create a new list
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     // Create a new list and return the new list back to the user
     // The list information will be passed in the JSON request body
     let newList = new List({
-        title: req.body.title
+        title: req.body.title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         res.send(listDoc)
@@ -101,9 +125,9 @@ app.post('/lists', (req, res) => {
 })
 
 // Update a specified list
-app.patch('/lists/:listId', (req, res) => {
+app.patch('/lists/:listId', authenticate, (req, res) => {
     // Update the specifiled list (list document with id in the URL) with the new values in the JSON body of the request
-    List.findOneAndUpdate({_id: req.params.listId}, {
+    List.findOneAndUpdate({_id: req.params.listId, _userId: req.user_id}, {
         $set: req.body
     }).then(() => {
         res.sendStatus(200);
@@ -111,15 +135,16 @@ app.patch('/lists/:listId', (req, res) => {
 });
 
 // Delete a list
-app.delete('/lists/:listId', (req, res) => {
+app.delete('/lists/:listId', authenticate, (req, res) => {
     // Delete the specified list (document with id in the URL)
-    List.findOneAndRemove({_id: req.params.listId}).then((removedListDoc) => {
-        res.send(removedListDoc)
+    List.findOneAndRemove({_id: req.params.listId, _userId: req.user_id}).then((removedListDoc) => {
+        res.send(removedListDoc);
+        deleteTasksFromList(removedListDoc._id);
     });
 });
 
 // Get all tasks inside a list
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     Task.find({
         _listId: req.params.listId
     }).then((tasks) => {
@@ -137,34 +162,88 @@ app.get('/lists/:listId/tasks/:taskId', (req, res) => {
     });
 })
 // Create a new task in a list 
-app.post('/lists/:listId/tasks', (req, res) => {
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    });
-    newTask.save().then((newTaskDoc) => {
-        res.send(newTaskDoc);
-    });
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object is found
+            return true;
+        }
+        //list object is not found
+        return false;
+    }).then((canCreateTask)=>{
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((newTaskDoc) => {
+                res.send(newTaskDoc);
+            });
+        } else {
+            res.sendStatus(404);
+        }
+    })
+    
 });
 
 // Update an existing task
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, { $set: req.body }).then(() => {
-        res.send({message: 'updated successfully!'})
-    });
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object is found
+            return true;
+        }
+        //list object is not found
+        return false;
+    }).then((canUpdateTask) => {
+        if(canUpdateTask) {
+            // if the currently authenticated user can update task
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, { $set: req.body }).then(() => {
+                res.send({message: 'updated successfully!'})
+            });
+        } else {
+            res.sendStatus(404);
+        }
+    })
+    
 });
 
 // Delete an existing task
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOneAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
-    });
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object is found
+            return true;
+        }
+        //list object is not found
+        return false;
+    }).then((canDeleteTask) => {
+        if (canDeleteTask){
+            Task.findOneAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            });
+        } else {
+            res.sendStatus(404);
+        }
+        
+    })
+    
 });
 
 // User routes
@@ -228,6 +307,15 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
     });
 });
     
+
+// Helper methods
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log("Tasks from " + _listId + " were deleted")
+    })
+}
 
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
